% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Unnest results from the output of pmap_safely}
\usage{
get_results(pmap_safely_output)
}
\arguments{
\item{pmap_safely_output}{The data frame outputted by the function pmap_safely_output. More specifically, must have an error and a result column (which is a list containing named lists or vectors).}
}
\description{
Allows for results to be easily analyzed.
}
\examples{
# a function to apply
calculate_if_positive <- function(a, b){
    if(a < 0 & b < 0) {stop("Both numbers are negative.")}
    else if(a < 0) {stop("Just the first number is negative")}
    else if(b < 0) {stop("Just the second number is negative")}

    list(add = a + b,
         subtract = a - b,
         multiply = a * b,
         divide = a / b)
}
# data frame to apply the function to by row
numbers <- data_frame(a = c(-1, 0, 1, 2),
                      b = c(2, 1, 0, -1),
                      irrelevant = c("minneapolis", "st_paul", "minneapolis", "st_paul"))

# apply and get results
output <- pmap_safely(numbers, calculate_if_positive)

get_results(output)
}
