% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_safely.R
\name{pmap_safely}
\alias{pmap_safely}
\title{Safely apply a function to each row of a data frame.}
\usage{
pmap_safely(.d, .f)
}
\arguments{
\item{.d}{A data frame with the inputs to .f.}

\item{.f}{A function to apply to each row of .d.}
}
\description{
Easily keep track of inputs and outputs of a function safely applied to each row of a data frame. Output is rectangularized into the original data frame for easy analysis.
}
\examples{
# a function to apply
calculate_if_positive <- function(a, b){
    if(a < 0 & b < 0) {stop("Both numbers are negative.")}
    else if(a < 0) {stop("Just the first number is negative")}
    else if(b < 0) {stop("Just the second number is negative")}

    list(add = a + b,
         subtract = a - b,
         multiply = a * b,
         divide = a / b)
}
# data frame to apply the function to by row
numbers <- data_frame(a = c(-1, 0, 1, 2),
                      b = c(2, 1, 0, -1),
                      irrelevant = c("minneapolis", "st_paul", "minneapolis", "st_paul"))

# apply
pmap_safely(numbers, calculate_if_positive)
}
