% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_errors.R
\name{get_errors_one_var}
\alias{get_errors_one_var}
\title{Get errors from pmap_safely_output for a specified variable.}
\usage{
get_errors_one_var(pmap_safely_output, var, specific = FALSE)
}
\arguments{
\item{pmap_safely_output}{The data frame outputted by the function pmap_safely_output. More specifically, must have an error and a result column (which is a list containing named lists or vectors).}

\item{var}{The input variable to analyzed errors based on.}

\item{specific}{Do you just want a summary or specific errors?}
}
\description{
Allows for errors to be easily analyzed.
}
