% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Unnest results from the output of pmap_safely}
\usage{
get_results(pmap_safely_output)
}
\arguments{
\item{pmap_safely_output}{The data frame outputted by the function pmap_safely_output. More specifically, must have an error and a result column (which is a list containing named lists or vectors).}
}
\description{
Allows for results to be easily analyzed.
}
