% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_safely.R
\name{pmap_safely}
\alias{pmap_safely}
\title{Safely apply a function to each row of a data frame.}
\usage{
pmap_safely(.d, .f)
}
\arguments{
\item{.d}{A data frame with the inputs to .f.}

\item{.f}{A function to apply to each row of .d.}
}
\description{
Easily keep track of inputs and outputs of a function safely applied to each row of a data frame. Output is rectangularized into the original data frame for easy analysis.
}
